+++ packages/typespec-go/src/tcgcadapter/types.ts
@@ -6,6 +6,7 @@
 import * as tcgc from '@azure-tools/typespec-client-generator-core';
 import * as tsp from '@typespec/compiler';
 import * as http from '@typespec/http';
+import * as helpers from './helpers.js';
 import * as go from '../../../codemodel.go/src/index.js';
 import * as naming from '../../../naming.go/src/naming.js';
 import { AdapterError } from '../tcgcadapter/errors.js';
@@ -71,7 +72,7 @@ export class TypeAdapter {
         continue;
       }
 
-      if (isPolymorphicRoot(modelType)) {
+      if (helpers.isPolymorphicRoot(modelType)) {
         // this is a root discriminated type
         const iface = this.getInterfaceType(modelType);
         this.getPkg().interfaces.push(iface);
@@ -111,7 +112,7 @@ export class TypeAdapter {
       }
       if (content.addlProps) {
         const annotations = new go.ModelFieldAnnotations(false, false, true, false);
-        const addlPropsType = new go.Map(this.getWireType(content.addlProps, false, false), isTypePassedByValue(content.addlProps));
+        const addlPropsType = new go.Map(this.getWireType(content.addlProps, false, false), helpers.isTypePassedByValue(content.addlProps));
         const addlProps = new go.ModelField('AdditionalProperties', addlPropsType, true, '', annotations);
         modelType.go.fields.push(addlProps);
       }
@@ -187,7 +188,7 @@ export class TypeAdapter {
           nullable = true;
         }
         // prefer elementTypeByValue. if false, then if the array elements have been explicitly marked as nullable then prefer that, else fall back to our usual algorithm
-        const myElementTypeByValue = elementTypeByValue ? true : nullable ? false : this.codeModel.options.sliceElementsByval || isTypePassedByValue(elementType);
+        const myElementTypeByValue = elementTypeByValue ? true : nullable ? false : this.codeModel.options.sliceElementsByval || helpers.isTypePassedByValue(elementType);
         const keyName = recursiveKeyName(`array-${myElementTypeByValue}`, elementType, substituteDiscriminator);
         let arrayType = this.types.get(keyName);
         if (arrayType) {
@@ -217,7 +218,7 @@ export class TypeAdapter {
       case 'utcDateTime':
         return this.getTimeType(type.encode, true);
       case 'dict': {
-        const valueTypeByValue = isTypePassedByValue(type.valueType);
+        const valueTypeByValue = helpers.isTypePassedByValue(type.valueType);
         const keyName = recursiveKeyName(`dict-${valueTypeByValue}`, type.valueType, substituteDiscriminator);
         let mapType = this.types.get(keyName);
         if (mapType) {
@@ -241,7 +242,7 @@ export class TypeAdapter {
         }
       }
       case 'model':
-        if (isPolymorphicRoot(type) && substituteDiscriminator) {
+        if (helpers.isPolymorphicRoot(type) && substituteDiscriminator) {
           return this.getInterfaceType(type);
         }
         return this.getModel(type);
@@ -463,7 +464,7 @@ export class TypeAdapter {
     if (model.name.length === 0) {
       throw new AdapterError('InternalError', 'unnamed model');
     }
-    if (!isPolymorphicRoot(model)) {
+    if (!helpers.isPolymorphicRoot(model)) {
       throw new AdapterError('InternalError', `type ${model.name} isn't a discriminator root`, model.__raw?.node);
     }
     let ifaceName = naming.createPolymorphicInterfaceName(naming.ensureNameCase(model.name));
@@ -487,7 +488,7 @@ export class TypeAdapter {
       throw new AdapterError('InternalError', `failed to find discriminator field for type ${model.name}`);
     }
     iface = new go.Interface(this.getPkg(), ifaceName, discriminatorField);
-    if (model.baseModel && isPolymorphicRoot(model.baseModel)) {
+    if (model.baseModel && helpers.isPolymorphicRoot(model.baseModel)) {
       iface.parent = this.getInterfaceType(model.baseModel);
     }
     this.types.set(ifaceName, iface);
@@ -513,19 +514,36 @@ export class TypeAdapter {
       usage |= go.UsageFlags.Output;
     }
 
-    const annotations = new go.ModelAnnotations(false, <tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.MultipartFormData) === tcgc.UsageFlags.MultipartFormData);
-    if (isPolymorphicRoot(model) || model.discriminatorValue) {
+    let omitSerde = false;
+    const omitValue = helpers.getClientOption('omitSerdeMethods', model, this.ctx.program);
+    if (omitValue) {
+      switch (omitValue) {
+        case 'all':
+          omitSerde = true;
+          break;
+        default:
+          this.ctx.program.reportDiagnostic({
+            code: 'InvalidClientOption',
+            severity: 'warning',
+            message: `invalid omitSerdeMethods value ${omitValue} on model ${model.name}`,
+            target: model.__raw?.node ?? tsp.NoTarget,
+          });
+      }
+    }
+
+    const annotations = new go.ModelAnnotations(omitSerde, <tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.MultipartFormData) === tcgc.UsageFlags.MultipartFormData);
+    if (helpers.isPolymorphicRoot(model) || model.discriminatorValue) {
       let iface: go.Interface | undefined;
       let discriminatorLiteral: go.Literal | undefined;
 
-      if (isPolymorphicRoot(model)) {
+      if (helpers.isPolymorphicRoot(model)) {
         // root type, we can get the InterfaceType directly from it
         iface = this.getInterfaceType(model);
       } else {
         // walk the parents until we find the first root type
         let parent = model.baseModel;
         while (parent) {
-          if (isPolymorphicRoot(parent)) {
+          if (helpers.isPolymorphicRoot(parent)) {
             iface = this.getInterfaceType(parent);
             break;
           }
@@ -551,7 +569,7 @@ export class TypeAdapter {
     } else {
       modelType = new go.Model(this.getPkg(), modelName, annotations, usage);
       // polymorphic types don't have XMLInfo
-      modelType.xml = adaptXMLInfo(this.getPkg(), model.decorators);
+      modelType.xml = helpers.adaptXMLInfo(this.getPkg(), model.decorators);
     }
 
     modelType.docs.summary = model.summary;
@@ -585,7 +603,7 @@ export class TypeAdapter {
     // for multipart/form data containing models, default to fields not being pointer-to-type as we
     // don't have to deal with JSON patch shenanigans. only the optional fields will be pointer-to-type.
     const isMultipartFormData = <tcgc.UsageFlags>(modelType.usage & tcgc.UsageFlags.MultipartFormData) === tcgc.UsageFlags.MultipartFormData;
-    let fieldByValue = isMultipartFormData ? true : isTypePassedByValue(prop.type);
+    let fieldByValue = isMultipartFormData ? true : helpers.isTypePassedByValue(prop.type);
     if (isMultipartFormData && prop.kind === 'property' && prop.optional) {
       fieldByValue = false;
     }
@@ -614,7 +632,7 @@ export class TypeAdapter {
       field.defaultValue = this.getDiscriminatorLiteral(prop);
     }
 
-    field.xml = adaptXMLInfo(this.getPkg(), prop.decorators, field);
+    field.xml = helpers.adaptXMLInfo(this.getPkg(), prop.decorators, field);
 
     return field;
   }
@@ -837,20 +855,6 @@ function recursiveKeyName(root: string, obj: tcgc.SdkType, substituteDiscriminat
   }
 }
 
-/**
- * returns true if the specified type doesn't need to be pointer-to-type
- * because it's implicitly nil-able.
- *
- * @param type the type to inspect
- * @returns true if the type is implicitly nil-able
- */
-export function isTypePassedByValue(type: tcgc.SdkType): boolean {
-  if (type.kind === 'nullable') {
-    type = type.type;
-  }
-  return type.kind === 'unknown' || type.kind === 'array' || type.kind === 'bytes' || type.kind === 'dict' || (type.kind === 'model' && isPolymorphicRoot(type));
-}
-
 interface ModelTypeSdkModelType {
   go: go.Model | go.PolymorphicModel;
   tcgc: tcgc.SdkModelType;
@@ -894,94 +898,4 @@ function aggregateProperties(sdkContext: tcgc.SdkContext, model: tcgc.SdkModelTy
     parent = parent.baseModel;
   }
   return { props: allProps, addlProps: addlProps };
-}
-
-// called for models and model fields. for the former, the field param will be undefined
-export function adaptXMLInfo(pkg: go.PackageContent, decorators: Array<tcgc.DecoratorInfo>, field?: go.ModelField): go.XMLInfo | undefined {
-  // if there are no decorators and this isn't a slice
-  // type in a model field then do nothing
-  if (decorators.length === 0 && (!field || field.type.kind !== 'slice')) {
-    return undefined;
-  }
-
-  const xmlInfo = new go.XMLInfo();
-  if (field && field.type.kind === 'slice') {
-    // for tsp, arrays are wrapped by default
-    xmlInfo.wraps = go.getTypeDeclaration(field.type.elementType, pkg);
-  }
-
-  const handleName = (decorator: tcgc.DecoratorInfo): void => {
-    if (field) {
-      xmlInfo.name = <string>decorator.arguments['name'];
-    } else {
-      // when applied to a model, it means the model's XML element
-      // node has a different name than the model.
-      xmlInfo.wrapper = <string>decorator.arguments['name'];
-    }
-  };
-
-  for (const decorator of decorators) {
-    switch (decorator.name) {
-      case 'TypeSpec.@encodedName':
-        if (decorator.arguments['mimeType'] === 'application/xml') {
-          handleName(decorator);
-        }
-        break;
-      case 'TypeSpec.Xml.@attribute':
-        xmlInfo.attribute = true;
-        break;
-      case 'TypeSpec.Xml.@name':
-        handleName(decorator);
-        break;
-      case 'TypeSpec.Xml.@unwrapped':
-        // unwrapped can only be applied to fields
-        if (field) {
-          switch (field.type.kind) {
-            case 'slice':
-              // unwrapped slice. default to using the serialized name
-              xmlInfo.wraps = undefined;
-              xmlInfo.name = field.serializedName;
-              break;
-            case 'string':
-              // an unwrapped string means it's text
-              xmlInfo.text = true;
-              break;
-          }
-        }
-        break;
-    }
-  }
-
-  return xmlInfo;
-}
-
-/**
- * returns any XMLInfo available for the provided type or undefined
- *
- * @param type the type to inspect for XMLInfo
- * @returns the XMLInfo or undefined
- */
-export function hasXMLInfo(type: go.WireType): go.XMLInfo | undefined {
-  if ('xml' in type) {
-    return type.xml;
-  }
-  return undefined;
-}
-
-/**
- * returns true if model is a polymorphic root type.
- *
- * @param model the model to inspect
- * @returns true if the model is a polymorphic root
- */
-function isPolymorphicRoot(model: tcgc.SdkModelType): boolean {
-  if (model.discriminatedSubtypes) {
-    // when there are sub-types we know for sure it's a polymorphic root
-    return true;
-  } else if (model.discriminatorProperty && !model.discriminatorValue) {
-    // we can land here if it's a root but has no child types
-    return true;
-  } else {
-    return false;
-  }
 }